Fix rust tests
Having a build.rs file enforces the crate to be build whenever tests are
launched at the workspace root (using the CI's `cargo test --all--features`
command for example).
However, in CI configuration, the variables PYTHON* and PYO3_PATH are
not set, thus the tests aren't launched because the build script panics.

To solve this, we delete the build.rs file and move its content to
src/main.rs, which is bit less "idiomatic", but functionally equivalent.